requirements.txt
flask~=2.1.0
flask_login~=0.6.0
python-dotenv~=0.20.0
cf_rules~=1.2.0
